Move tasks our of taskgroups and remove unused code
