+++ dlme_airflow/services/harvest_dag_generator.py
@@ -8,14 +8,7 @@ from airflow.operators.dummy import DummyOperator
 from airflow.models import Variable
 
 # Our stuff
-from task_groups.validate_dlme_metadata import build_validate_metadata_taskgroup, build_sync_metadata_taskgroup
-from task_groups.harvest import build_harvester_taskgroup
-from task_groups.post_harvest import build_post_havest_taskgroup
-from task_groups.detect_metadata_changes import build_detect_metadata_changes_taskgroup
-from task_groups.transform import build_transform_taskgroup
-from task_groups.index import build_index_taskgroup
-from task_groups.harvest_report import build_harvest_report_taskgroup
-from task_groups.send_harvest_report import build_send_harvest_report_taskgroup
+from task_groups.validate_dlme_metadata import build_validate_metadata_taskgroup
 from task_groups.etl import build_provider_etl_taskgroup
 from utils.driver_tag import fetch_driver
 
@@ -45,14 +38,9 @@ def create_dag(provider, default_args):
     with dag:
         validate_dlme_metadata = build_validate_metadata_taskgroup(provider, dag)
 
-        # harvester = build_harvester_taskgroup(provider, dag)
-
         # A dummy operator is required as a transition point between task groups
         harvest_begin = DummyOperator(task_id='harvest_begin', trigger_rule='none_failed', dag=dag)
         harvest_complete = DummyOperator(task_id='harvest_complete', trigger_rule='none_failed', dag=dag)
-        # transform_complete = DummyOperator(task_id='transform_complete', trigger_rule='none_failed', dag=dag)
-        # index_complete = DummyOperator(task_id='index_complete', trigger_rule='none_failed', dag=dag)
-        # report_complete = DummyOperator(task_id='report_complete', trigger_rule='none_failed', dag=dag)
 
         # TODO
         # post_harvest = build_post_havest_taskgroup(provider, dag)
@@ -60,22 +48,7 @@ def create_dag(provider, default_args):
         # TODO: A dummy operator is required as a transition point between task groups
         # post_harvest_complete = DummyOperator(task_id='post_harvest_complete', trigger_rule='none_failed', dag=dag)
 
-        # sync_dlme_metadata = build_sync_metadata_taskgroup(provider, dag)
-        # TODO
-        # collect_metadata_changes = build_detect_metadata_changes_taskgroup(provider, dag)
-
-        # transform = build_transform_taskgroup(provider, dag)
-        # index = build_index_taskgroup(provider, dag)
-        # report = build_harvest_report_taskgroup(provider, dag)
-        # send_report = build_send_harvest_report_taskgroup(provider, dag)
-
         etl = build_provider_etl_taskgroup(provider, dag)
-
-        # # validate_dlme_metadata >> harvester >> harvest_complete >> post_harvest >> post_harvest_complete >> collect_metadata_changes
-        # # validate_dlme_metadata >> harvester >> harvest_complete >> collect_metadata_changes
-        # validate_dlme_metadata >> harvester >> harvest_complete
-        # harvest_complete >> sync_dlme_metadata >> transform >> transform_complete
-        # transform_complete >> index >> index_complete >> report >> report_complete >> send_report
 
         validate_dlme_metadata >> harvest_begin >> etl >> harvest_complete
