+++ dlme_airflow/tasks/transform.py
@@ -39,26 +39,3 @@ def build_transform_task(coll_label, data_path, task_group: TaskGroup, dag: DAG)
         },
         dag=dag
     )
- 
-
-def transform_tasks(provider, task_group: TaskGroup, dag: DAG) -> TaskGroup:
-    task_array = []
-    source = catalog_for_provider(provider)
-
-    try:
-        collections = iter(list(source))
-        for collection in collections:
-            coll_label = f"{provider}.{collection}"
-            data_path = catalog_for_provider(coll_label).metadata.get('data_path', f"{provider}/{collection}")
-            task_array.append(build_transform_task(coll_label, data_path, task_group, dag))
-    except TypeError:
-        data_path = source.metadata.get('data_path', f"{provider}")
-        return build_transform_task(f"{provider}", data_path, task_group, dag)
-
-    return task_array
-
-
-def build_transform_taskgroup(provider, dag: DAG) -> TaskGroup:
-    transform_taskgroup = TaskGroup(group_id="transform")
-
-    return transform_tasks(provider, transform_taskgroup, dag)
